Cache updated Cache class set to abstract. FIFO, Belady, LRU, MRU, Random Cache classed added with minimal implementations added
